+++ Engine/Containers/Freelist.hpp
@@ -3,8 +3,8 @@
 #include "Defines.hpp"
 
 struct FreelistNode {
-	unsigned long long offset;
-	unsigned long long size;
+	unsigned long offset;
+	unsigned long size;
 	struct FreelistNode* next = nullptr;
 };
 
@@ -18,7 +18,7 @@ public:
 	* 
 	* @param total_size The total size in bytes that the free list should track.
 	*/
-	DAPI void Create(unsigned long long total_size);
+	DAPI void Create(unsigned long total_size);
 
 	/*
 	* @brief Destroys list.
@@ -32,7 +32,7 @@ public:
 	* @param offset A pointer to hold the offset to the allocated memory.
 	* @return bool True if a block of memory has found and allocated; otherwise false.
 	*/
-	DAPI bool AllocateBlock(unsigned long long size, unsigned long long* offset);
+	DAPI bool AllocateBlock(unsigned long size, unsigned long * offset);
 
 	/*
 	* @brief Attempts to free a free block of memory at the given offset, and of the 
@@ -42,14 +42,14 @@ public:
 	* @param offset A pointer to hold the offset to the allocated memory.
 	* @return bool True if a block of memory has free; otherwise false.
 	*/
-	DAPI bool FreeBlock(unsigned long long size, unsigned long long offset);
+	DAPI bool FreeBlock(unsigned long size, unsigned long long offset);
 
 	/**
 	 * @brief Attempts to resize the freelist
 	 * 
 	 * @param new_size The new size of memory the freelist could hold.
 	 */
-	DAPI bool Resize(unsigned long long new_size);
+	DAPI bool Resize(unsigned long new_size);
 
 	/*
 	* @brief Clears the free list.
@@ -60,15 +60,15 @@ public:
 	* @brief Returns the amount of free space in this list.
 	* NOTE: Since this has to iterate the entire internal list, this can be an expensive operation. Use sparingly.
 	*/
-	DAPI unsigned long long GetFreeSpace();
+	DAPI unsigned long GetFreeSpace();
 
 private:
 	FreelistNode* AcquireFreeNode();
 	void ResetNode(FreelistNode* node);
 
 private:
-	unsigned long long TotalSize;
-	unsigned long long MaxEntries;
+	unsigned long TotalSize;
+	unsigned long MaxEntries;
 	FreelistNode* Head;
 	FreelistNode* Nodes;
